Changes to try and support a single hardware Serial board
These are changes to try and support a board with just one hardwareSerial device, rather than the two of the Uno Wifi R2.

In 485test.ino, line 5, SERIAL_PORT_HARDWARE is defined as Serial1. This could be redefined to Serial, which compiles and runs but doesn't work. This value is passed as a pointer on startup and everywhere uses the pointer now.

Also, throughout the code, Serial is no longer used as the debugging output. Rather I have created a macro, DEBUGOUT(), which allows you to define the output device as Serial, a LCD, or none. This is necessary because if you're using Serial for the RS485 it can't be used for output.

This requires the updated RS485.h and .cpp, which are also available in my repository.

In ZLFP10Thermostat.h are defines that define where the debug output goes.
#define DEBUGOUTPUTDEVICE_SERIAL to send it to Serial,
#define DEBUGOUTPUTDEVICE_LCD to send it to LCD,
undefine DEBUGOUTPUT_ENABLED  to send it nowhere.