Modified to use SoftwareSerial and ModbusMaster
Modified for boards that don't have a second HardwareSerial -- like the original Uno -- or where the HardwareSerial doesn't work, like the R4 Uno Wifi.

In order to do this I had to switch to using the SoftwareSerial, which meant switching to a different Modbus library, ModbusMaster, because ArduinoModbus only works on HardwareSerial.

I had to modify ModbusMaster, and that modification is included.